+++ realm/realm-library/src/main/cpp/io_realm_internal_objectstore_OsApp.cpp
@@ -155,15 +155,14 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_objectstore_OsApp_nativeCreate(JN
             client_config.custom_encryption_key = encryption_key.transform<std::vector<char>>();
         }
 
-        // FIXME: SyncManager is still a singleton. Should be refactored to allow multiple
-        SyncManager::shared().configure(client_config, app_config);
+        SharedApp app = App::get_shared_app(app_config, client_config);
         // Init logger. Must be called after .configure()
-        SyncManager::shared().set_logger_factory(s_sync_logger_factory);
+        app->sync_manager()->set_logger_factory(s_sync_logger_factory);
         // Register Sync Client thread start/stop callback. Must be called after .configure()
         static AndroidClientListener client_thread_listener(env);
         g_binding_callback_thread_observer = &client_thread_listener;
 
-        return reinterpret_cast<jlong>(new std::shared_ptr<App>(SyncManager::shared().app()));
+        return reinterpret_cast<jlong>(new std::shared_ptr<App>(app));
     }
     CATCH_STD()
     return 0;