+++ realm/realm-library/src/main/cpp/io_realm_mongodb_sync_ClientResetRequiredError.cpp
@@ -16,6 +16,7 @@
 
 #include <jni.h>
 
+#include <sync/app.hpp>
 #include <sync/sync_manager.hpp>
 
 #include "util.hpp"
@@ -23,12 +24,13 @@
 
 using namespace realm;
 
-JNIEXPORT void JNICALL Java_io_realm_mongodb_sync_ClientResetRequiredError_nativeExecuteClientReset(JNIEnv* env, jobject,
+JNIEXPORT void JNICALL Java_io_realm_mongodb_sync_ClientResetRequiredError_nativeExecuteClientReset(JNIEnv* env, jlong j_app_ptr, jobject,
                                                                                jstring localRealmPath)
 {
     try {
+        auto app = *reinterpret_cast<std::shared_ptr<app::App>*>(j_app_ptr);
         JStringAccessor local_realm_path(env, localRealmPath);
-        if (!SyncManager::shared().immediately_run_file_actions(std::string(local_realm_path))) {
+        if (!app->sync_manager()->immediately_run_file_actions(std::string(local_realm_path))) {
             ThrowException(
                 env, IllegalState,
                 concat_stringdata("Realm was not configured correctly. Client Reset could not be run for Realm at: ",